Johannes,
Based on someone else's suggestion
(http://bugs.gentoo.org/show_bug.cgi?id=168709), I've updated colorgcc
to be language aware. However, I don't know perl. Do you know perl well
enough to clean up what I've done? Even if you don't, are you willing to
host the upgraded colorgcc on your site (once it is cleaned up)?

The modifications to the perl script are as follows:
Replace the regex that looks for "warning" with a variable. That
variable's value is retrieved from a file based on the current language
setting. If the language key is not found, a quick compile is done to
determine the value and it is appended to the lookup file.

Example:

LANG=en_us

in warnings_db.txt:
en_us=warning

It still needs some work to be more portable and to facilitate
installers. I've included the new function below.

Thanks,

Harvey